admin chnage

+++ database/migrations/2022_04_17_201244_add_fields_to_donations_table.php
@@ -16,7 +16,7 @@ class AddFieldsToDonationsTable extends Migration
         Schema::table('donations', function (Blueprint $table) {
             //
             $table->string('type');
-            $table->integer('amount');
+            $table->double('amount');
             $table->string('first_name');
             $table->string('last_name');
             $table->string('gender')->nullable();
